+++ src/acconeer/exptool/a121/algo/distance/_detector_plugin.py
@@ -26,6 +26,7 @@ from acconeer.exptool.a121.algo._plugins import (
 from acconeer.exptool.app.new import (
     BUTTON_ICON_COLOR,
     AppModel,
+    BackendLogger,
     ConnectionState,
     GeneralMessage,
     HandledException,
@@ -58,9 +59,6 @@ from ._detector import (
 )
 
 
-log = logging.getLogger(__name__)
-
-
 @attrs.mutable(kw_only=True)
 class SharedState:
     sensor_id: int = attrs.field(default=1)
@@ -81,6 +79,7 @@ class BackendPlugin(DetectorBackendPluginBase[SharedState]):
         self._recorder: Optional[a121.H5Recorder] = None
         self._opened_record: Optional[a121.H5Record] = None
         self._detector_instance: Optional[Detector] = None
+        self._log = BackendLogger.getLogger(__name__)
 
         self.restore_defaults()
 
@@ -143,7 +142,7 @@ class BackendPlugin(DetectorBackendPluginBase[SharedState]):
                 context_group = f.create_group("context")
                 self.shared_state.context.to_h5(context_group)
         except Exception:
-            log.warning("Detector could not write to cache")
+            self._log.warning("Detector could not write to cache")
 
         self.detach_client()
 
@@ -488,6 +487,7 @@ class ViewPlugin(DetectorViewPluginBase):
     def __init__(self, app_model: AppModel, view_widget: QWidget) -> None:
         super().__init__(app_model=app_model, view_widget=view_widget)
         self.app_model = app_model
+        self._log = logging.getLogger(__name__)
 
         sticky_layout = QVBoxLayout()
         sticky_layout.setContentsMargins(0, 0, 0, 0)
@@ -711,7 +711,7 @@ class ViewPlugin(DetectorViewPluginBase):
     # TODO: move to detector base (?)
     def handle_message(self, message: GeneralMessage) -> None:
         if message.name == "sync":
-            log.debug(f"{type(self).__name__} syncing")
+            self._log.debug(f"{type(self).__name__} syncing")
 
             self.config_editor.sync()
         else: