+++ src/acconeer/exptool/app/new/__init__.py
@@ -11,7 +11,14 @@ from ._enums import (
 from ._exceptions import HandledException
 from .app import main
 from .app_model import AppModel
-from .backend import BackendPlugin, GeneralMessage, Message, PluginStateMessage, is_task
+from .backend import (
+    BackendLogger,
+    BackendPlugin,
+    GeneralMessage,
+    Message,
+    PluginStateMessage,
+    is_task,
+)
 from .pluginbase import PlotPluginBase, PluginSpecBase, ViewPluginBase
 from .storage import get_temp_dir, get_temp_h5_path
 from .ui import BUTTON_ICON_COLOR