+++ src/acconeer/exptool/app/new/backend/__init__.py
@@ -2,11 +2,13 @@
 # All rights reserved
 
 from ._backend import Backend, ClosedTask
+from ._backend_logger import BackendLogger
 from ._backend_plugin import BackendPlugin
 from ._message import (
     BackendPluginStateMessage,
     ConnectionStateMessage,
     GeneralMessage,
+    LogMessage,
     Message,
     PluginStateMessage,
     StatusMessage,