+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 - Serial port tagging for XE123, XE124, XE125 modules
 - Disable connect button for an unflashed XC120 device
+- Silent backend logging on Windows
 
 ### Added
 - Backend CPU load in status bar.